feature(separete into different components, better implementation of passing props)
